pytorch_ver/requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cycler==0.11.0
decorator==5.1.1
fonttools==4.29.1
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.38.0
matplotlib==3.5.1
numba==0.55.1
numpy==1.19.0
packaging==21.3
Pillow==9.0.1
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
resampy==0.2.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==3.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.8
